[FIX] Integrate Dashboard and Pop Up Notification
